Revert "fix: checks type in helper functions (#53)"
This reverts commit 5e3c780f67ce1311c3190dc911f56a0c2cb98fea.
